Revert "Actu query3"
This reverts commit c22102342917f965ba5d503415dc059ef974f470.

+++ src/main/java/pe/edu/upc/waba/dtos/ComentarioDTO.java
@@ -9,7 +9,7 @@ import pe.edu.upc.waba.entities.Profesor;
 
 public class ComentarioDTO {
     private int idCo;
-    private double puntosCo;
+    private float puntosCo;
     private String resenaCo;
     private Profesor Pro;
 
@@ -21,7 +21,7 @@ public class ComentarioDTO {
         this.idCo = idCo;
     }
 
-    public double getPuntosCo() {
+    public float getPuntosCo() {
         return puntosCo;
     }
